Remove lib/resque/core_ext/kernel.rb and instead put it in test_helper.rb
