--- lib/resque/core_ext/kernel.rb
@@ -1,16 +1,0 @@
-module Kernel
-  def silence_warnings
-    with_warnings(nil) { yield }
-  end unless Kernel.respond_to?(:silence_warnings)
-
-  def with_warnings(flag)
-    old_verbose, $VERBOSE = $VERBOSE, flag
-    yield
-  ensure
-    $VERBOSE = old_verbose
-  end unless Kernel.respond_to?(:with_warnings)
-
-  def jruby?
-    defined?(JRUBY_VERSION)
-  end
-end
