+++ test/legacy/test_helper.rb
@@ -3,7 +3,6 @@ require 'timeout'
 require 'bundler/setup'
 require 'redis/namespace'
 require 'minitest/autorun'
-require 'resque/core_ext/kernel'
 
 $dir = File.dirname(File.expand_path(__FILE__))
 $LOAD_PATH.unshift $dir + '/../lib'
@@ -77,6 +76,27 @@ module PerformJob
   def perform_job(klass, *args)
     resque_job = Resque::Job.new(:testqueue, 'class' => klass, 'args' => args)
     resque_job.perform
+  end
+end
+
+
+##
+# Add helper methods to Kernel
+#
+module Kernel
+  def silence_warnings
+    with_warnings(nil) { yield }
+  end unless Kernel.respond_to?(:silence_warnings)
+
+  def with_warnings(flag)
+    old_verbose, $VERBOSE = $VERBOSE, flag
+    yield
+  ensure
+    $VERBOSE = old_verbose
+  end unless Kernel.respond_to?(:with_warnings)
+
+  def jruby?
+    defined?(JRUBY_VERSION)
   end
 end
